Try to find an exact match before doing a regex match.
I was trying to use the import feature of gcalcli when it kept telling
me that I did not have a single calendar for the command. When you give
a calendar on the command line with the --cal option it does a regex
match to find your Google calendar. That seems OK until you have two
calendars w/ the same word in it. This commit modifies the search to try
to do an exact string comparison first before doing the regex match.
There are probably side effects to this too, but I think this is a
better way of doing things